taking out logic inside handleAddBook func in a single function on utils file
